+++ CHANGELOG.md
@@ -42,6 +42,9 @@
   [#4783](https://github.com/wasm-bindgen/wasm-bindgen/pull/4783)
   [#4782](https://github.com/wasm-bindgen/wasm-bindgen/pull/4782)
 
+* Fixed a multithreaded wasm32+atomics race where `Atomics.waitAsync` promise callbacks could call `run` without waking first, causing sporadic panics.
+  [#TODO](https://github.com/wasm-bindgen/wasm-bindgen/pull/xxxx)
+
 ### Removed
 
 ## [0.2.105](https://github.com/wasm-bindgen/wasm-bindgen/compare/0.2.104...0.2.105)